Rename gem to sidekiq-alive-next
