+++ spec/spec_helper.rb
@@ -1,5 +1,5 @@
 require 'bundler/setup'
-require 'sidekiq-alive'
+require 'sidekiq-alive-next'
 require 'rspec-sidekiq'
 require 'mock_redis'
 require 'rack/test'
